% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_colors.R
\docType{data}
\name{gg_colors}
\alias{gg_colors}
\title{Provide color support for ggplots.}
\format{
An object of class \code{character} of length 14.
}
\usage{
gg_colors
}
\value{
colors as a vector
}
\description{
Provide color support for ggplots.
}
\keyword{datasets}
