% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_lpirfs_lin.R
\name{gg_lpirfs_lin}
\alias{gg_lpirfs_lin}
\title{Plot impulse response of linear local projections in lpirfs package.}
\usage{
gg_lpirfs_lin(
  est_result,
  chart_title = glue::glue("{est_result$specs$endog_data} to {est_result$specs$shock}"),
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{est_result}{estimation results from the lpirfs package}

\item{chart_title}{The chart title}

\item{xlab}{label on x axis}

\item{ylab}{label on y axis}
}
\value{
Returns a ggplot2 chart
}
\description{
Plot impulse response of linear local projections in lpirfs package.
}
