% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_theme_irf.R
\name{gg_theme_irf}
\alias{gg_theme_irf}
\title{This function applies a custom theme to ggplot2 plots.}
\usage{
gg_theme_irf()
}
\value{
A ggplot2 theme object.
}
\description{
This function applies a custom theme to ggplot2 plots.
}
