% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_vars.R
\name{gg_vars}
\alias{gg_vars}
\title{Plot impulse response from vars package.}
\usage{
gg_vars(
  irf_result,
  var_shock,
  var_endo = var_irf$response,
  line_color = "black",
  shade_color = "#009CDE",
  alpha_trans = 0.5,
  scale_option = "free",
  xlab = NULL,
  ylab = NULL,
  num_col = NULL,
  return_only_data = FALSE
)
}
\arguments{
\item{irf_result}{impulse response from a var}

\item{var_shock}{Which shock in the VAR}

\item{var_endo}{Which endogenous variables to plot? Default is all variables}

\item{line_color}{Provide the line color of impulse response (mean value)}

\item{shade_color}{Provide the shade color of impulse response (confidence interval)}

\item{alpha_trans}{Transparency of shade colors}

\item{scale_option}{Provide scaling options. Options are free_x, free_y and free. For details refer to ggplot2 facet_wrap}

\item{xlab}{label on x axis}

\item{ylab}{label on y axis}

\item{num_col}{Number of columns when faceting charts}

\item{return_only_data}{Should return only the resulting data instead of ggplot2 chart? Useful for post processing and comparison with other methods}
}
\value{
Returns a ggplot2 chart
}
\description{
Plot impulse response from vars package.
}
